refactor: extract dialog code to ModMain.Dialog.cs partial class
ModMain.cs: removed ~650 lines of dialog code, added 'partial' keyword.
ModMain.Dialog.cs: new file with all dialog logic (MyMsgBoxConverter,
MyMsgBox*, MyMsgBoxTick, Dialog_OnShow, CreateDialogFromConverter, etc.).
Zero call-site changes — all public API remains on ModMain.
